requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.0.3
feedparser==6.0.10
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
APScheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
loguru==0.7.2
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
python-dateutil==2.8.2
nest-asyncio==1.5.8